feat(binance): add spot public websocket streams (#8)
## Summary
- add Binance spot public websocket stream support for last price,
partial order book, trades, and klines
- map SDK websocket payloads into mkt MarketDataEvent values with
bounded buffering and terminal error/close handling
- add explicit async stream close() semantics and document deterministic
shutdown

## Review
- Linus self-review round 1 found unbounded buffering and weak drop-time
shutdown; fixed with bounded queue and terminal overflow handling
- Linus self-review round 2 found non-terminal SDK errors/closes and
incomplete deterministic shutdown; fixed with terminal messages and
EventStream::close()
- final narrow Linus verification found no blocking/medium lifecycle
findings

## Tests
- scripts/check-format-and-lint.sh
- cargo test --workspace

---------

+++ crates/mkt-core/src/traits.rs
@@ -20,7 +20,7 @@ pub trait MarketData: Send + Sync {
     async fn markets(&self) -> Result<Vec<MarketInfo>>;
     async fn last_price(&self, symbol: &Symbol) -> Result<LastPrice>;
     async fn last_prices(&self, symbols: Option<&[Symbol]>) -> Result<Vec<LastPrice>>;
-    async fn order_book(&self, symbol: &Symbol, depth: Option<u32>) -> Result<OrderBook>;
+    async fn order_book(&self, symbol: &Symbol, depth: Option<u16>) -> Result<OrderBook>;
     async fn recent_trades(&self, symbol: &Symbol, limit: Option<u32>) -> Result<Vec<Trade>>;
     async fn klines(&self, request: KlineRequest) -> Result<Vec<Kline>>;
 }
